# api.yaml
openapi: 3.0.3
info:
  title: chat server
  description: this server does provide a api for chat functionality
  version: 1.0.0
servers:
  - url: 'https:chat.schlagh.com'
components:
  securitySchemes:
    authorization:
      type: authorization
      name: authorization
      in: header
      description: the json web token
...
paths:
...
  /message/:
    put:
      description: |
        add a new message
      summary: add new message
      operationId: addMessage
      parameters:
        - $ref: '#components/securitySchemes/authorization'
        - name: body
          in: header
          required: true
          schema:
            $ref: '#components/schemas/MessageInput'
      responses:
        '200':
          description: successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  mid:
                    type: number
                    description: the id of the new message
        '400':
          description: bad request
        '403':
          $ref: '#components/responses/NotAllowed'
        '500':
          description: internal server error
  /message/add:
    put:
      description: |
        add a new message
      summary: add new message
      operationId: addMessage
      parameters:
        - $ref: '#components/securitySchemes/authorization'
        - name: body
          in: header
          required: true
          schema:
            type: object
            properties:
              chatType:
                type: string
                description: |
                  the type of the chat,
                  either normalChat or groupChat
              chatId:
                type: number
                description: the id of the chat
              message:
                $ref: '#components/schemas/MessageInput'
      responses:
        '200':
          description: successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  mid:
                    type: number
                    description: the id of the new message
        '400':
          description: bad request
        '403':
          $ref: '#components/responses/NotAllowed'
        '500':
          description: internal server error
  /message/load:
    post:
      description: |
        load messages
      summary: load messages
      operationId: loadMessages
      parameters:
      - $ref: '#components/securitySchemes/authorization'
      - name: body
        in: header
        required: true
        schema:
          type: object
          properties:
            chatType:
              type: string
              description: the type of the chat
            chatId:
              type: number
              description: the id of the chat where the messages should be loaded
            lastMsgId:
              type: number
              description: the id of the last message that has been loaded
            num:
              type: number
              description: how many messages should be loaded?
      responses:
        '200':
          description: successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    description: |
                      the status of this request
                        reached top:
                          all messages have been loaded, there could be some in the body
                        success:
                          messages loaded, there are some left who are not loaded
                  messages:
                    type: array
                    description: the messages
                    items:
                      $ref: '#components/schemas/Message'
        '400':
          description: bad request
        '403':
          $ref: '#components/responses/NotAllowed'
        '500':
          description: internal server error
...